requirements/base.txt
Django==1.5.4
South==0.8.2
argparse==1.2.1
boto==2.13.3
distribute==0.7.3
django-bootstrap3==0.0.7
django-haystack==2.1.0
django-storages==1.1.8
django-markupfield==1.2.0
django-taggit==0.10
django-widget-tweaks==1.3
oauthlib==0.6.0
psycopg2==2.5.1
pysolr==3.1.0
python-openid==2.2.5
python-social-auth==0.1.13
pytz==2013d
requests==2.0.0
requests-oauthlib==0.3.3
six==1.4.1
wsgiref==0.1.2
Markdown==2.3.1
